vim-patch:9.2.0689: the "%" command is slow on a long line with many slashes
Problem:  The "%" command can be very slow on a long line that contains
          many slashes, for example a line of base64 data.
Solution: When looking for a line comment, scan the line only once while
          skipping over strings, instead of rescanning from the start for
          every slash.  Move check_linecomment() to cindent.c so it can
          reuse the file-local skip_string().

related: vim/vim#20491
fixes:   vim/vim#20557
closes:  vim/vim#20575

https://github.com/vim/vim/commit/9f9af034adcdb5773e19c4bb52e3c368bbce826e
