+++ src/nvim/search.c
@@ -2068,10 +2068,11 @@ pos_T *findmatchlimit(oparg_T *oap, int initc, int flags, int64_t maxtravel)
       }
     }
 
-    // Track block comment state when FM_SKIPCOMM is set.
+    // Track block comment state when FM_SKIPCOMM is set.  Markers inside a
+    // string are not comments, so skip them while "inquote" is set.
     // Backward: '/' of end-marker enters comment; '*' of start-marker exits.
     // Forward:  '/' of start-marker enters comment; '/' of end-marker exits.
-    if (skip_comments && !comment_dir) {
+    if (skip_comments && !comment_dir && !inquote) {
       if (backwards) {
         // Guard pos.col < comment_col: don't misread '* /' at the '//'
         // position as a block-comment end-marker.
@@ -2337,56 +2338,6 @@ pos_T *findmatchlimit(oparg_T *oap, int initc, int flags, int64_t maxtravel)
     return &pos;
   }
   return (pos_T *)NULL;         // never found it
-}
-
-/// Check if line[] contains a / / comment.
-/// @returns MAXCOL if not, otherwise return the column.
-int check_linecomment(const char *line)
-{
-  const char *p = line;  // scan from start
-  // skip Lispish one-line comments
-  if (curbuf->b_p_lisp) {
-    if (vim_strchr(p, ';') != NULL) {   // there may be comments
-      bool in_str = false;       // inside of string
-
-      while ((p = strpbrk(p, "\";")) != NULL) {
-        if (*p == '"') {
-          if (in_str) {
-            if (*(p - 1) != '\\') {             // skip escaped quote
-              in_str = false;
-            }
-          } else if (p == line || ((p - line) >= 2
-                                   // skip #\" form
-                                   && *(p - 1) != '\\' && *(p - 2) != '#')) {
-            in_str = true;
-          }
-        } else if (!in_str && ((p - line) < 2
-                               || (*(p - 1) != '\\' && *(p - 2) != '#'))
-                   && !is_pos_in_string(line, (colnr_T)(p - line))) {
-          break;                // found!
-        }
-        p++;
-      }
-    } else {
-      p = NULL;
-    }
-  } else {
-    while ((p = vim_strchr(p, '/')) != NULL) {
-      // Accept a double /, unless it's preceded with * and followed by *,
-      // because * / / * is an end and start of a C comment.  Only
-      // accept the position if it is not inside a string.
-      if (p[1] == '/' && (p == line || p[-1] != '*' || p[2] != '*')
-          && !is_pos_in_string(line, (colnr_T)(p - line))) {
-        break;
-      }
-      p++;
-    }
-  }
-
-  if (p == NULL) {
-    return MAXCOL;
-  }
-  return (int)(p - line);
 }
 
 /// Move cursor briefly to character matching the one under the cursor.
